Delete APIProvider stub, divide it to smaller service stubs.
